fix: added bold versions of LinkIcon plus fixed file names and a couple of bugs in the d-prop for <p
