+++ packages/icons-react/src/icons/link/LinkIcon.tsx
@@ -1,14 +1,16 @@
 import { type FC } from "react";
 import { makeIconComponent } from "../../IconFactory";
 import { type IconProps } from "../types";
-import { LinkMedium } from "./LinkIconMedium";
-import { LinkSmall } from "./LinkIconSmall";
+import { LinkMedium } from "./LinkMedium";
+import { LinkMediumBold } from "./LinkMediumBold";
+import { LinkSmall } from "./LinkSmall";
+import { LinkSmallBold } from "./LinkSmallBold";
 
 export const LinkIcon: FC<IconProps> = makeIconComponent({
     bold: {
-        small: LinkSmall,
-        medium: LinkMedium,
-        inherit: LinkMedium,
+        small: LinkSmallBold,
+        medium: LinkMediumBold,
+        inherit: LinkMediumBold,
     },
     regular: {
         small: LinkSmall,
