+++ packages/icons-react/src/icons/link/LinkMediumBold.tsx
@@ -1,0 +1,12 @@
+import React, { FC, SVGAttributes } from "react";
+
+export const LinkMediumBold: FC<SVGAttributes<SVGElement>> = (props) => (
+    <svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M17.4697 6.53033C16.0304 5.09107 14.0067 5.05398 12.9377 6.12293L10.5303 8.53034L9.46966 7.46968L11.8771 5.06227C13.6718 3.26751 16.6559 3.59522 18.5303 5.46967C20.4048 7.34412 20.7325 10.3282 18.9377 12.1229L16.5303 14.5303L15.4697 13.4697L17.8771 11.0623C18.946 9.99332 18.9089 7.96959 17.4697 6.53033ZM15.5304 9.53033L9.53035 15.5303L8.46969 14.4697L14.4697 8.46967L15.5304 9.53033ZM8.53033 10.5303L6.12292 12.9377C5.05398 14.0067 5.09107 16.0304 6.53033 17.4697C7.96959 18.9089 9.99332 18.946 11.0623 17.8771L13.4697 15.4697L14.5303 16.5303L12.1229 18.9377C10.3282 20.7325 7.34412 20.4048 5.46967 18.5303C3.59522 16.6559 3.26751 13.6718 5.06226 11.8771L7.46967 9.46967L8.53033 10.5303Z"
+            fill="currentColor"
+        />
+    </svg>
+);
