+++ packages/icons-react/src/icons/link/LinkSmall.tsx
@@ -1,0 +1,12 @@
+import React, { FC, SVGAttributes } from "react";
+
+export const LinkSmall: FC<SVGAttributes<SVGElement>> = (props) => (
+    <svg viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M14.3131 5.68688C13.1695 4.54328 11.5422 4.49829 10.6704 5.37001L8.798 7.24244L8.09089 6.53533L9.96332 4.6629C11.3189 3.3073 13.5865 3.54605 15.0202 4.97978C16.454 6.41351 16.6927 8.68108 15.3371 10.0367L13.4647 11.9091L12.7576 11.202L14.63 9.32958C15.5017 8.45785 15.4567 6.83049 14.3131 5.68688ZM12.6869 8.02021L8.02022 12.6869L7.31311 11.9798L11.9798 7.31311L12.6869 8.02021ZM7.24244 8.79799L5.37001 10.6704C4.49829 11.5421 4.54328 13.1695 5.68688 14.3131C6.83049 15.4567 8.45785 15.5017 9.32958 14.63L11.202 12.7576L11.9091 13.4647L10.0367 15.3371C8.68108 16.6927 6.41351 16.454 4.97978 15.0202C3.54605 13.5865 3.3073 11.3189 4.6629 9.96332L6.53533 8.09089L7.24244 8.79799Z"
+            fill="currentColor"
+        />
+    </svg>
+);
