+++ packages/icons-react/src/icons/link/LinkSmallBold.tsx
@@ -1,0 +1,12 @@
+import React, { FC, SVGAttributes } from "react";
+
+export const LinkSmallBold: FC<SVGAttributes<SVGElement>> = (props) => (
+    <svg viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M14.1363 5.86367C13.0653 4.7926 11.598 4.79604 10.8472 5.5468L8.97477 7.41923L7.91411 6.35857L9.78654 4.48614C11.2631 3.00957 13.6907 3.29675 15.197 4.80301C16.7033 6.30928 16.9904 8.73691 15.5139 10.2135L13.6414 12.0859L12.5808 11.0252L14.4532 9.15281C15.204 8.40206 15.2074 6.93474 14.1363 5.86367ZM12.8637 8.197L8.197 12.8637L7.13634 11.803L11.803 7.13634L12.8637 8.197ZM7.41922 8.97478L5.54679 10.8472C4.79603 11.598 4.79259 13.0653 5.86366 14.1364C6.93473 15.2074 8.40204 15.204 9.1528 14.4532L11.0252 12.5808L12.0859 13.6415L10.2135 15.5139C8.73689 16.9905 6.30926 16.7033 4.803 15.197C3.29674 13.6908 3.00956 11.2631 4.48613 9.78655L6.35856 7.91412L7.41922 8.97478Z"
+            fill="currentColor"
+        />
+    </svg>
+);
